Added first draft of sqlite data model backend

+++ SatisfactoryCalculator/Source/ApplicationServices/CalculationService/FuelCalculationResult.cs
@@ -4,10 +4,10 @@ namespace SatisfactoryCalculator.Source.ApplicationServices;
 internal class FuelCalculationResult
 {
     // ReSharper disable once UnusedAutoPropertyAccessor.Global
-    public double FuelBurnTime { get; set; }
-    public double PowerProduction { get; set; }
-    public double AmountPerMinute { get; set; }
-    public double? SupplementalAmountPerMinute { get; set; }
-    public double? ByProductAmountPerMinute { get; set; }
+    public decimal FuelBurnTime { get; set; }
+    public decimal PowerProduction { get; set; }
+    public decimal AmountPerMinute { get; set; }
+    public decimal? SupplementalAmountPerMinute { get; set; }
+    public decimal? ByProductAmountPerMinute { get; set; }
 }
