+++ SatisfactoryCalculator/Source/ApplicationServices/CalculationService/ICalculationService.cs
@@ -6,8 +6,8 @@ namespace SatisfactoryCalculator.Source.ApplicationServices;
 
 internal interface ICalculationService
 {
-    double? CalculateAmountPerMinte(Form? form, double? amount, double manufactoringDuration);
-    double NormalizeAmount(Form? form, double sourceInput);
+    decimal? CalculateAmountPerMinte(Form? form, decimal? amount, decimal manufactoringDuration);
+    decimal NormalizeAmount(Form? form, decimal sourceInput);
     FuelCalculationResult CalculateFuelConsumption(GeneratorFuel generatorFuel, double overclock);
     RecipeItemProductionResult CalculateRecipeItemProduction(IRecipe recipe, IEntity entity, IBuilding building, double overclock);
     RecipeBuildingProductionResult CalculateRecipeBuildingProduction(IRecipe recipe, IBuilding building, double overclock);
