+++ SatisfactoryCalculator/Source/ApplicationServices/CalculationService/RecipeItemProductionResult.cs
@@ -3,9 +3,9 @@ namespace SatisfactoryCalculator.Source.ApplicationServices;
 
 internal class RecipeItemProductionResult
 {
-    public double Amount { get; init; }
-    public double AmountPerMinute { get; init; }
-    public double PowerConsumption { get; init; }
+    public decimal Amount { get; init; }
+    public decimal AmountPerMinute { get; init; }
+    public decimal PowerConsumption { get; init; }
     // ReSharper disable once UnusedAutoPropertyAccessor.Global
-    public double Time { get; set; }
+    public decimal Time { get; set; }
 }
