+++ SatisfactoryCalculator/Source/Models/Implementation/Buildings/Building.cs
@@ -7,9 +7,9 @@ internal class Building : IBuilding
     public string ClassName { get; init; }
     public string Name { get; init; }
     public string Description { get; init; }
-    public double? ManufactoringSpeed { get; init; }
-    public double? PowerConsumption { get; init; }
-    public double? PowerConsumptionExponent { get; init; }
+    public decimal? ManufactoringSpeed { get; init; }
+    public decimal? PowerConsumption { get; init; }
+    public decimal? PowerConsumptionExponent { get; init; }
     public PowerConsumptionRange? PowerConsumptionRange { get; init; }
     public BitmapImage? Image { get; init; }
 }
