+++ SatisfactoryCalculator/Source/Models/Implementation/Buildings/GeneratorFuel/Generator.cs
@@ -3,7 +3,7 @@ namespace SatisfactoryCalculator.Source.Models;
 
 internal class Generator : Building, IGenerator
 {
-    public double PowerProduction { get; init; }
-    public double? SupplementalToPowerRatio { get; init; }
-    public double? SupplementalLoadAmount { get; init; }
+    public decimal PowerProduction { get; init; }
+    public decimal? SupplementalToPowerRatio { get; init; }
+    public decimal? SupplementalLoadAmount { get; init; }
 }
