+++ SatisfactoryCalculator/Source/Models/Implementation/Recipe/Recipe.cs
@@ -13,7 +13,7 @@ internal class Recipe : IRecipe
     public bool ConstructedByBuildGun { get; init; }
     public bool ConstructedInWorkshop { get; init; }
     public bool ConstructedInWorkbench { get; init; }
-    public double ManufactoringDuration { get; init; }
+    public decimal ManufactoringDuration { get; init; }
     public RecipeBuilding[] Buildings { get; init; }
     public RecipePart[] Ingredients { get; init; }
     public RecipePart[] Products { get; init; }
