+++ SatisfactoryCalculator/Source/Models/Interfaces/IAmmunition.cs
@@ -2,9 +2,9 @@ namespace SatisfactoryCalculator.Source.Models.Interfaces;
 
 internal interface IAmmunition : IItem
 {
-    public double MaxAmmoEffectiveRange { get; init; }
-    public double? ReloadTimeMultiplier { get; init; }
-    public double FireRate { get; init; }
-    public double WeaponDamageMultiplier { get; init; }
+    public decimal MaxAmmoEffectiveRange { get; init; }
+    public decimal? ReloadTimeMultiplier { get; init; }
+    public decimal FireRate { get; init; }
+    public decimal WeaponDamageMultiplier { get; init; }
     public IWeapon UsedInWeapon { get; set; }
 }
