+++ SatisfactoryCalculator/Source/Models/Interfaces/IConsumable.cs
@@ -2,5 +2,5 @@ namespace SatisfactoryCalculator.Source.Models.Interfaces;
 
 internal interface IConsumable : IItem
 {
-    public double? HealthGain { get; init; }
+    public decimal? HealthGain { get; init; }
 }
