+++ SatisfactoryCalculator/Source/Models/Interfaces/IGenerator.cs
@@ -2,7 +2,7 @@ namespace SatisfactoryCalculator.Source.Models.Interfaces;
 
 internal interface IGenerator : IBuilding
 {
-    double PowerProduction { get; }
-    double? SupplementalToPowerRatio { get; }
-    double? SupplementalLoadAmount { get; }
+    decimal PowerProduction { get; }
+    decimal? SupplementalToPowerRatio { get; }
+    decimal? SupplementalLoadAmount { get; }
 }
