+++ SatisfactoryCalculator/Source/Models/Interfaces/IWeapon.cs
@@ -2,9 +2,9 @@ namespace SatisfactoryCalculator.Source.Models.Interfaces;
 
 internal interface IWeapon : IItem
 {
-    double? DamageMultiplier { get; init; }
-    double? ReloadTime { get; init; }
-    double? AutoReloadDelay { get; init; }
+    decimal? DamageMultiplier { get; init; }
+    decimal? ReloadTime { get; init; }
+    decimal? AutoReloadDelay { get; init; }
     EquipmentSlot EquipmentSlot { get; init; }
     public IAmmunition[] Ammunitions { get; set; }
 }
