+++ Shared/Utility/PowerConsumptionRangeConverterUtility.cs
@@ -2,16 +2,16 @@ namespace SatisfactoryCalculator.Shared.Utility;
 
 public static class PowerConsumptionRangeConverterUtility
 {
-	public static PowerConsumptionRange? ConverToPowerConsumption(double? estimatedMininumPowerConsumption, double? estimatedMaximumPowerConsumption)
+	public static PowerConsumptionRange? ConverToPowerConsumption(decimal? estimatedMininumPowerConsumption, decimal? estimatedMaximumPowerConsumption)
 	{
 		return estimatedMininumPowerConsumption.HasValue
 			? new PowerConsumptionRange(estimatedMininumPowerConsumption.Value,estimatedMaximumPowerConsumption!.Value)
 			: null;
 	}
 
-	public static PowerConsumptionRange? ConverToPowerVariableConsumption(double variablePowerConsumptionConstant, double variablePowerConsumptionFactor)
+	public static PowerConsumptionRange? ConverToPowerVariableConsumption(decimal variablePowerConsumptionConstant, decimal variablePowerConsumptionFactor)
 	{
-		return variablePowerConsumptionConstant != 0.0
+		return variablePowerConsumptionConstant != (decimal)0.0
 			? new PowerConsumptionRange(variablePowerConsumptionConstant, variablePowerConsumptionConstant + variablePowerConsumptionFactor)
 			: null;
 	}
